// packages/v-uibuilder/src/enums/settingEnum.ts
export enum ChangeSetterEnum {
  ALL = "all",
  ATTRIBUTE = "attribute",
  STYLE = "style",
  EVENT = "event",
  MOVEABLE = "moveable",
}

export enum ModelAttributeEnum {
  INIT_STYLE = "initStyle",
  ON_DOM_EVENT = "onDomEvent",
  ON_MESSAGE_META = "onMessageMeta",
  ON_WATCH_SETTING = "onWatchSetting",
  LIFE_CYCLE = "lifeCycle",
}

export enum EndToEndEnum {
  WORK_SPACE = "workspace",
  EDIT_PAGE = "editPage",
  ADD_PAGE = "addPage",
  PREVIEW_PAGE = "previewPage",
  IMPORT_APP = "importApp",
  PREVIEW_WEBSITE = "previewWebsite",
  EDIT_WEBSITE = "editWebsite",
  ADD_WEBSITE = "addWebsite",
}

export enum LoginTypeEnum {
  UIB = "uib",
  EMEI = "emei",
}

export enum recoverTypesEnum {
  ROUTER = "router",
  ATTR_BIND = "attr_bind",
}