// packages/v-uibuilder/src/enums/pageEnum.ts
export enum PageEnum {
  // basic login path
  BASE_LOGIN = "/login",
  // basic home path
  BASE_HOME = "/management/flowChart",
  // error page path
  ERROR_PAGE = "/exception",
  // error log page path
  ERROR_LOG_PAGE = "/error-log/list",
  // 管理页面(用于调试)
  MANAGEMENT = "/uibuilder/management",
  // 配置页面
  SETTING = "/uibuilder/setting",
  //工作空间
  WORKSPACE = "/uibuilder/workspace",
  //编辑页面
  EDIT = "/uibuilder/edit",
  //编辑页面
  WEBSITE = "/uibuilder/website",
  // 发布页
  PUBLISH = "/uibuilder/publish"
}